+++ src/components/groups/group.entity.ts
